+++ web/src/components/pages/mainPage/components/mainPageCarousel/style.css
@@ -1,0 +1,39 @@
+.mainPageCarouselItem__container {
+    position: relative;
+    width: 50%;
+}
+
+.mainPageCarouselItem__img {
+    margin-top: 40px;
+    opacity: 1;
+    display: block;
+    max-width: 350px;
+    height: auto;
+    transition: 0.5s ease;
+    backface-visibility: hidden;
+}
+
+.mainPageCarouselItem__txt__container {
+    transition: 0.5s ease;
+    opacity: 0;
+    position: absolute;
+    top: 50%;
+    left: 80%;
+    transform: translate(-50%, -50%);
+    -ms-transform: translate(-50%, -50%);
+    text-align: center;
+    width: 12rem;
+}
+
+.mainPageCarouselItem__container:hover .mainPageCarouselItem__img {
+    opacity: 0.3;
+}
+
+.mainPageCarouselItem__container:hover .mainPageCarouselItem__txt__container {
+    opacity: 1;
+}
+
+.mainPageCarouselItem__txt {
+    font-size: 20px;
+    color: #04aa6d;
+}
