+++ web/src/components/pages/mainPage/components/welcome/style.css
